+++ tests-ui/tests/renderer/thumbnail/composables/useWorkflowThumbnail.spec.ts
@@ -6,7 +6,7 @@ import {
   useWorkflowStore
 } from '@/platform/workflow/management/stores/workflowStore'
 
-vi.mock('@/renderer/thumbnail/graphThumbnailRenderer', () => ({
+vi.mock('@/renderer/core/thumbnail/graphThumbnailRenderer', () => ({
   createGraphThumbnail: vi.fn()
 }))
 
@@ -22,10 +22,10 @@ vi.mock('@/scripts/api', () => ({
 }))
 
 const { useWorkflowThumbnail } = await import(
-  '@/renderer/thumbnail/useWorkflowThumbnail'
+  '@/renderer/core/thumbnail/useWorkflowThumbnail'
 )
 const { createGraphThumbnail } = await import(
-  '@/renderer/thumbnail/graphThumbnailRenderer'
+  '@/renderer/core/thumbnail/graphThumbnailRenderer'
 )
 const { api } = await import('@/scripts/api')
